+++ mcp/sse_test.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"sync/atomic"
 	"testing"
 
@@ -85,95 +84,6 @@ func TestSSEServer(t *testing.T) {
 			} else {
 				ss.Close()
 				cs.Wait()
-			}
-		})
-	}
-}
-
-func TestScanEvents(t *testing.T) {
-	tests := []struct {
-		name    string
-		input   string
-		want    []event
-		wantErr string
-	}{
-		{
-			name:  "simple event",
-			input: "event: message\nid: 1\ndata: hello\n\n",
-			want: []event{
-				{name: "message", id: "1", data: []byte("hello")},
-			},
-		},
-		{
-			name:  "multiple data lines",
-			input: "data: line 1\ndata: line 2\n\n",
-			want: []event{
-				{data: []byte("line 1\nline 2")},
-			},
-		},
-		{
-			name:  "multiple events",
-			input: "data: first\n\nevent: second\ndata: second\n\n",
-			want: []event{
-				{data: []byte("first")},
-				{name: "second", data: []byte("second")},
-			},
-		},
-		{
-			name:  "no trailing newline",
-			input: "data: hello",
-			want: []event{
-				{data: []byte("hello")},
-			},
-		},
-		{
-			name:    "malformed line",
-			input:   "invalid line\n\n",
-			wantErr: "malformed line",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := strings.NewReader(tt.input)
-			var got []event
-			var err error
-			for e, err2 := range scanEvents(r) {
-				if err2 != nil {
-					err = err2
-					break
-				}
-				got = append(got, e)
-			}
-
-			if tt.wantErr != "" {
-				if err == nil {
-					t.Fatalf("scanEvents() got nil error, want error containing %q", tt.wantErr)
-				}
-				if !strings.Contains(err.Error(), tt.wantErr) {
-					t.Fatalf("scanEvents() error = %q, want containing %q", err, tt.wantErr)
-				}
-				return
-			}
-
-			if err != nil {
-				t.Fatalf("scanEvents() returned unexpected error: %v", err)
-			}
-
-			if len(got) != len(tt.want) {
-				t.Fatalf("scanEvents() got %d events, want %d", len(got), len(tt.want))
-			}
-
-			for i := range got {
-				if g, w := got[i].name, tt.want[i].name; g != w {
-					t.Errorf("event %d: name = %q, want %q", i, g, w)
-				}
-				if g, w := got[i].id, tt.want[i].id; g != w {
-					t.Errorf("event %d: id = %q, want %q", i, g, w)
-				}
-				if g, w := string(got[i].data), string(tt.want[i].data); g != w {
-					t.Errorf("event %d: data = %q, want %q", i, g, w)
-				}
 			}
 		})
 	}
